backend/requirements.txt
fastapi[standard]
uvicorn
websockets
azure-cognitiveservices-speech 
python-dotenv 
asyncio
aiohttp